Added file names, added warehouses and items. Changed all the files so conform to same style.
