+++ CSharpAPI/Controllers/ItemTypesController.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Mvc;
 namespace CSharpAPI.Controllers
 {
     [ApiController]
-    [Route("api/v1/[controller]")]
+    [Route("api/v1/itemtypes")]
     public class ItemTypesController : ControllerBase
     {
         private readonly IItemTypeService _service;
