+++ CSharpAPI/Controllers/LocationsController.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Mvc;
 namespace CSharpAPI.Controllers
 {
     [ApiController]
-    [Route("api/v1/[controller]")]
+    [Route("api/v1/locations")]
     public class LocationsController : ControllerBase
     {
         private readonly ILocationService _service;
