+++ CSharpAPI/Program.cs
@@ -1,4 +1,5 @@
 using CSharpAPI.Service;
+using CSharpAPI.Services;
 using Microsoft.OpenApi.Models;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -12,9 +13,15 @@ builder.Services.AddSingleton<IWarehouseService, WarehouseService>();
 builder.Services.AddSingleton<IItemTypeService, ItemTypeService>();
 builder.Services.AddSingleton<IItemGroupService, ItemGroupService>();
 builder.Services.AddSingleton<IItemLineService, ItemLineService>();
-// builder.Services.AddSingleton<ITransfersService, TransferSerivce>();
+builder.Services.AddSingleton<ITransfersService, TransferSerivce>();
 builder.Services.AddSingleton<IShipmentService, ShipmentService>();
 builder.Services.AddSingleton<ILocationService, LocationService>();
+builder.Services.AddSingleton<IItemsService, ItemsService>();
+builder.Services.AddSingleton<ISupplierService, SupplierService>();
+builder.Services.AddSingleton<IInventoriesService, InventoriesService>();
+builder.Services.AddSingleton<IClientsService, ClientsService>();
+
+
 
 // Add CORS
 builder.Services.AddCors(options =>
