+++ CSharpAPI/data/itemtypes.json
@@ -54,5 +54,19 @@
     "description": "Test Type Description",
     "created_at": "2024-11-07T21:39:33.684652Z",
     "updated_at": "2024-11-07T21:39:33.684652Z"
+  },
+  {
+    "id": 11,
+    "name": "Test Type",
+    "description": "Test Type Description",
+    "created_at": "2024-11-11T10:58:11.37804Z",
+    "updated_at": "2024-11-11T10:58:11.37804Z"
+  },
+  {
+    "id": 12,
+    "name": "Test Type",
+    "description": "Test Type Description",
+    "created_at": "2024-11-11T10:58:12.297414Z",
+    "updated_at": "2024-11-11T10:58:12.297414Z"
   }
 ]
